+++ app/services/sortable-is-transformable.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-sortable/services/sortable-is-transformable';
