--- tests/dummy/app/components/sortable-item-ie-fix.js
@@ -1,5 +1,0 @@
-import SortableItem from 'ember-sortable/components/sortable-item';
-import SortableItemMixinIE from 'ember-sortable/mixins/sortable-item-ie-fix';
-
-export default SortableItem.extend(SortableItemMixinIE, {
-});
