--- tests/dummy/app/templates/components/sortable-item-ie-fix.hbs
@@ -1,1 +1,0 @@
-{{yield}}
